--- a/src/bitmap.h
+++ b/src/bitmap.h
@@ -586,4 +586,20 @@ inline Color Bitmap::GetShadowColor() const {
 	return sh_color;
 }
 
+inline int Bitmap::GetWidth() const {
+	return width();
+}
+
+inline int Bitmap::GetHeight() const {
+	return height();
+}
+
+inline Rect Bitmap::GetRect() const {
+	return Rect(0, 0, width(), height());
+}
+
+inline bool Bitmap::GetTransparent() const {
+	return format.alpha_type != PF::NoAlpha;
+}
+
 #endif