--- a/src/pending_message.cpp
+++ b/src/pending_message.cpp
@@ -83,7 +83,7 @@ std::string PendingMessage::ApplyTextInsertingCommands(std::string input, uint32
 	while (iter != end) {
 		auto ret = Utils::UTF8Next(iter, end);
 		if (ret.ch != escape_char) {
-			iter = ret.iter;
+			iter = ret.next;
 			continue;
 		}
 
@@ -95,7 +95,7 @@ std::string PendingMessage::ApplyTextInsertingCommands(std::string input, uint32
 		output.append(start_copy, iter);
 		start_copy = iter;
 
-		iter = ret.iter;
+		iter = ret.next;
 		if (iter == end) {
 			break;
 		}