--- a/src/text.cpp
+++ b/src/text.cpp
@@ -23,15 +23,33 @@
 #include "bitmap.h"
 #include "font.h"
 #include "text.h"
-#include "game_system.h"
+#include "compiler.h"
 
 #include <cctype>
 #include <iterator>
 
-void Text::Draw(Bitmap& dest, int x, int y, int color, FontRef font, std::string const& text, Text::Alignment align) {
-	if (text.length() == 0) return;
+Rect Text::Draw(Bitmap& dest, int x, int y, Font& font, const Bitmap& system, int color, char32_t ch, bool is_exfont) {
+	if (is_exfont) {
+		return Font::exfont->Render(dest, x, y, system, color, ch);
+	} else {
+		return font.Render(dest, x, y, system, color, ch);
+	}
+}
 
-	Rect dst_rect = font->GetSize(text);
+Rect Text::Draw(Bitmap& dest, int x, int y, Font& font, Color color, char32_t ch, bool is_exfont) {
+	if (is_exfont) {
+		return Font::exfont->Render(dest, x, y, color, ch);
+	} else {
+		return font.Render(dest, x, y, color, ch);
+	}
+}
+
+Rect Text::Draw(Bitmap& dest, const int x, const int y, Font& font, const Bitmap& system, const int color, const std::string& text, const Text::Alignment align) {
+	if (text.length() == 0) return { x, y, 0, 0 };
+
+	Rect dst_rect = font.GetSize(text);
+
+	const int ih = dst_rect.height;
 
 	switch (align) {
 	case Text::AlignCenter:
@@ -45,82 +63,69 @@ void Text::Draw(Bitmap& dest, int x, int y, int color, FontRef font, std::string
 
 	dst_rect.y = y;
 	dst_rect.width += 1; dst_rect.height += 1; // Need place for shadow
-	if (dst_rect.IsOutOfBounds(dest.GetWidth(), dest.GetHeight())) return;
+	if (dst_rect.IsOutOfBounds(dest.GetWidth(), dest.GetHeight())) return { x, y, 0, 0 };
 
-	BitmapRef text_surface; // Complete text will be on this surface
-	text_surface = Bitmap::Create(dst_rect.width, dst_rect.height, true);
-	text_surface->Clear();
-
-	BitmapRef system = Cache::SystemOrBlack();
+	const int iy = dst_rect.y;
+	const int ix = dst_rect.x;
 
 	// Where to draw the next glyph (x pos)
 	int next_glyph_pos = 0;
 
-	// The current char is an exfont
-	bool is_exfont = false;
-
 	// This loops always renders a single char, color blends it and then puts
 	// it onto the text_surface (including the drop shadow)
-	std::u32string u32text = Utils::DecodeUTF32(text);
-	for (auto c = u32text.begin(), end = u32text.end(); c != end; ++c) {
-		Rect next_glyph_rect(next_glyph_pos, 0, 0, 0);
+	auto iter = &*text.begin();
+	const auto end = &*text.end();
+	while (iter != end) {
+		auto ret = Utils::TextNext(iter, end, 0);
 
-		char32_t const next_c = std::distance(c, end) > 1? *std::next(c) : 0;
-
-		// ExFont-Detection: Check for A-Z or a-z behind the $
-		if (*c == '$' && std::isalpha(next_c)) {
-			int exfont_value = -1;
-			// Calculate which exfont shall be rendered
-			if (islower(next_c)) {
-				exfont_value = 26 + next_c - 'a';
-			} else if (isupper(next_c)) {
-				exfont_value = next_c - 'A';
-			} else { assert(false); }
-			is_exfont = true;
-
-			Font::exfont->Render(*text_surface, next_glyph_rect.x, next_glyph_rect.y, *system, color, exfont_value);
-		} else { // Not ExFont, draw normal text
-			font->Render(*text_surface, next_glyph_rect.x, next_glyph_rect.y, *system, color, *c);
-		}
-
-		// If it's a full size glyph, add the size of a half-size glyph twice
-		if (is_exfont) {
-			is_exfont = false;
-			next_glyph_pos += 12;
-			// Skip the next character
-			++c;
-		} else {
-			next_glyph_pos += font->GetSize(*c).width;
-		}
-	}
-
-	BitmapRef text_bmp = Bitmap::Create(*text_surface, text_surface->GetRect());
-
-	Rect src_rect(0, 0, dst_rect.width, dst_rect.height);
-	int iy = dst_rect.y;
-	if (dst_rect.height > text_bmp->GetHeight()) {
-		iy += ((dst_rect.height - text_bmp->GetHeight()) / 2);
-	}
-	int ix = dst_rect.x;
-
-	dest.Blit(ix, iy, *text_bmp, src_rect, 255);
-}
-
-void Text::Draw(Bitmap& dest, int x, int y, Color color, FontRef font, std::string const& text) {
-	if (text.length() == 0) return;
-
-	int next_glyph_pos = 0;
-
-	for (char32_t c : Utils::DecodeUTF32(text)) {
-		if (c == U'\n') {
-			y += font->GetSize(c).height;
-			next_glyph_pos = 0;
+		iter = ret.next;
+		if (EP_UNLIKELY(!ret)) {
 			continue;
 		}
-		Rect next_glyph_rect(x + next_glyph_pos, y, 0, 0);
-
-		font->Render(dest, next_glyph_rect.x, next_glyph_rect.y, color, c);
-
-		next_glyph_pos += font->GetSize(c).width;
+		next_glyph_pos += Text::Draw(dest, ix + next_glyph_pos, iy, font, system, color, ret.ch, ret.is_exfont).width;
 	}
+	return { x, y, next_glyph_pos, ih };
 }
+
+Rect Text::Draw(Bitmap& dest, const int x, const int y, Font& font, const Color color, const std::string& text) {
+	if (text.length() == 0) return { x, y, 0, 0 };
+
+	int dx = x;
+	int mx = x;
+
+	int dy = y;
+	int ny = 0;
+
+	auto iter = &*text.begin();
+	const auto end = &*text.end();
+	while (iter != end) {
+		auto ret = Utils::UTF8Next(iter, end);
+
+		iter = ret.next;
+		if (EP_UNLIKELY(!ret)) {
+			continue;
+		}
+
+		if (ret.ch == U'\n') {
+			if (ny == 0) {
+				ny = font.GetSize(ret.ch).height;
+			}
+			dy += ny;
+			mx = std::max(mx, dx);
+			dx = x;
+			ny = 0;
+			continue;
+		}
+
+		auto rect = font.Render(dest, dx, dy, color, ret.ch);
+		dx += rect.width;
+		assert(ny == 0 || ny == rect.height);
+		ny = rect.height;
+	}
+	dy += ny;
+	mx = std::max(mx, dx);
+
+	return Rect(x, y, mx - x , dy - y);
+}
+
+