… bench/font.cpp
… bench/text.cpp
… src/utils.h
… src/window_battlestatus.cpp
… src/window_message.cpp
… src/window_message.h
… tests/font.cpp
… tests/text.cpp
… tests/utf.cpp