~~Depend on: #1956~~

This PR refactors the underlying `Font` and `Text` interfaces to be more streamlined and efficient.

Current Test Results
--------------------------

After "Font::Render - remove query for system"

Font:
```
------------------------------------------------------
Benchmark               Time           CPU Iterations
------------------------------------------------------
BM_FontSizeStr        637 ns        637 ns    1107660
BM_Glyph              171 ns        171 ns    4005735
BM_Render            2481 ns       2480 ns     282438
```
Text:
```
---------------------------------------------------------
Benchmark                  Time           CPU Iterations
---------------------------------------------------------
BM_TextDrawStr         94971 ns      94971 ns       7365
BM_TextDrawStrRaw      42719 ns      42718 ns      16451
```

HEAD of PR:

Font:
```
-------------------------------------------------------
Benchmark                Time           CPU Iterations
-------------------------------------------------------
BM_FontSizeStr         636 ns        632 ns    1116692
BM_FontSizeChar         21 ns         21 ns   34673138
BM_Glyph                67 ns         67 ns   10619258
BM_Render             2182 ns       2182 ns     320334
```
Text:
```
---------------------------------------------------------------
Benchmark                        Time           CPU Iterations
---------------------------------------------------------------
BM_TextDrawStrSystem         86675 ns      86676 ns       7963
BM_TextDrawStrColor          41405 ns      41405 ns      17102
BM_TextDrawCharSystem         2210 ns       2211 ns     316856
BM_TextDrawCharSystemEx       3046 ns       3046 ns     230286
BM_TextDrawCharColor           965 ns        965 ns     733548
BM_TextDrawCharColorEx        1971 ns       1971 ns     355208
```