@fmatthew5876 2019-12-30
Here is a Vtune profile of Messages 4 vs this. The test is a map which repeatedly draws text each frame.

#1956 
![msg4](https://user-images.githubusercontent.com/1198466/71568827-29a07a80-2a98-11ea-8c52-7b4ec9d0b97c.png)

This PR
![msg5](https://user-images.githubusercontent.com/1198466/71568828-2f965b80-2a98-11ea-945b-f80b626034bd.png)


This PR is finished. Next will be another cleanup of `Window_Message`.