@fmatthew5876 2020-01-08
Addressed review comment and added tests to `Makefile.am`