@fmatthew5876 2020-01-20
This PR was tested pretty heavily in development of #2026 . It should be merged soon.